+++ apps/winn/src/winn_codegen_resolve.erl
@@ -51,6 +51,7 @@ resolve_dot_call('Protocol', Fun) -> {winn_protocol, Fun};
 resolve_dot_call('Health', Fun)   -> {winn_health, Fun};
 resolve_dot_call('Metrics', Fun)  -> {winn_metrics, Fun};
 resolve_dot_call('Agent', Fun)    -> {winn_agent, Fun};
+resolve_dot_call('Pipeline', Fun) -> {winn_pipeline, Fun};
 resolve_dot_call('ReplBindings', get) -> {winn_repl, get_binding};
 resolve_dot_call(Mod, Fun) ->
     ErlMod = list_to_atom(string:lowercase(atom_to_list(Mod))),